[infra] Cria bypass de memória como comportamento padrão para download (#902) (#1026)
* reorganize folders to support medata functions

* download now bypass memory and reauth is a lone function

* remove print and update pyproject.toml

* adds capability to download unlimited file sizes

* adds reauth and updates download

* delete prints
