+++ python-package/pyproject.toml
@@ -13,7 +13,7 @@ packages = [
 ]
 readme = "README.md"
 repository = "https://github.com/base-dos-dados/bases"
-version = "1.6.0"
+version = "1.6.1-alpha.0"
 
 [tool.poetry.scripts]
 basedosdados = 'basedosdados.cli.cli:cli'
